network call changed, styling format changed
